+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Update the available Accepted data types for Job profiles (UIDATIMP-561)
 * Update the available Incoming record options for Match profiles (UIDATIMP-562)
 * Update the available Data type options for File extension settings (UIDATIMP-626)
+* Connect MARC field protection settings to the server (UIDATIMP-621)
 * Field mappings: Item - update reference dropdown list for Item status (UIDATIMP-529)
 
 ### Bugs fixed: